requirements.txt
asgiref==3.6.0
Django==4.1.7
django-js-asset==2.0.0
django-simple-history==3.3.0
djangorestframework==3.14.0
psycopg2==2.9.5
pytz==2022.7.1
sqlparse==0.4.3
typing-extensions==4.5.0
tzdata==2022.7
